+++ src/main/java/com/yuzee/app/dao/IArticleDAO.java
@@ -21,7 +21,7 @@ public interface IArticleDAO {
 	Integer getTotalSearchCount(final Integer startIndex, final Integer pageSize, final String sortByField, final String sortByType, final String searchKeyword,
 			List<String> categoryIdList, List<String> tagList, Boolean status, Date date);
 
-	List<String> getAuthors(int startIndex, Integer pageSize, String searchString);
+	List<String> getAuthors(Long startIndex, Integer pageSize, String searchString);
 
 	int getTotalAuthorCount(String searchString);
